requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
pydantic-settings
pymupdf
pdf2image
pillow
chromadb
sentence-transformers
google-generativeai
langchain-community
python-dotenv
